% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_mtmegps_dependencies}
\alias{install_mtmegps_dependencies}
\title{Install Python dependencies for MTMEGPS}
\usage{
install_mtmegps_dependencies()
}
\value{
None
}
\description{
This function installs the required Python environment
for running deep learning models in MTMEGPS, including
TensorFlow and Keras3 through the reticulate interface.
}
